chore: move k8sutils to utils folder
