+++ .testcoverage.yml
@@ -59,7 +59,7 @@ exclude:
     - main
     - mocks
     - stubs
-    - ^internal/k8sutils # k8sutils has to be dissected into smaller single-purpose packages, for now excluding from coverage check
+    - ^internal/utils/k8s # k8sutils has to be dissected into smaller single-purpose packages, for now excluding from coverage check
     - internal/utils/logger
     - ^internal/reconciler/logparser # logparser reconciler will be removed soon
     - ^test # e2e and integration tests
