+++ internal/reconciler/logparser/sync_test.go
@@ -14,7 +14,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client/fake"
 
 	telemetryv1alpha1 "github.com/kyma-project/telemetry-manager/apis/telemetry/v1alpha1"
-	"github.com/kyma-project/telemetry-manager/internal/k8sutils/mocks"
+	"github.com/kyma-project/telemetry-manager/internal/utils/k8s/mocks"
 )
 
 var (
