+++ internal/utils/k8s/owner_reference_setter.go
@@ -1,4 +1,4 @@
-package k8sutils
+package k8s
 
 import (
 	"context"
